refactor :: sendQrEvent package 경로 변경
